Move Providers into components
