+++ context/app/static/js/components/App.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import Container from '@material-ui/core/Container';
-import Providers from '../Providers';
+import Providers from './Providers';
 import Routes from './Routes';
 import Footer from '../Footer';
 import Header from '../Header';
