+++ context/app/static/js/components/Providers.jsx
@@ -3,7 +3,7 @@ import { ThemeProvider } from 'styled-components';
 import PropTypes from 'prop-types';
 import { MuiThemeProvider, StylesProvider, createGenerateClassName } from '@material-ui/core/styles';
 import CssBaseline from '@material-ui/core/CssBaseline';
-import theme from './theme';
+import theme from '../theme';
 
 const generateClassName = createGenerateClassName({
   disableGlobal: true,
