+++ context/app/static/js/test-utils.js
@@ -1,7 +1,7 @@
 /* eslint-disable import/no-extraneous-dependencies */
 import React from 'react';
 import { render } from '@testing-library/react';
-import Providers from './Providers';
+import Providers from './components/Providers';
 
 const AllTheProviders = ({ children }) => {
   return <Providers>{children}</Providers>;
